pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paperdigest"
version = "0.1.0"
description = "Automated research paper digest system"
requires-python = ">=3.9"
dependencies = [
    "arxiv>=2.1",
    "requests>=2.31",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
llm = ["openai>=1.0"]
dev = ["pytest>=7.0"]

[project.scripts]
paperdigest = "paperdigest.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
